feat : set caffeine cache config

+++ src/main/kotlin/com/ruuniv/common/cache/CacheType.kt
@@ -3,7 +3,7 @@ package com.ruuniv.common.cache
 enum class CacheType(
     val cacheName: String,
     val expireAfterWrite: Long,
-    val maxSize: Int
+    val maxSize: Long
 ) {
     API_KEY("API_KEY", 3600, 10000),
     STUDENT("STUDENT", 3600, 10000),
